Implement encoding, decoding, and test for utf-8
Practically, a count leading-zeros function is first implement.
The clz function  is used in encoding. Besides, encoding function
incudes a exponent estimation process to accelerate the
calculation. Decoding function is implemented quite straight
forward. Finally, the test include two part. Reason for the two
test is because number does not need encode for those who is less
than 16. Both parts decode a number between 0 to 255 and then
compare the original number with the number after decoding and
encoding.